+++ js/main.js
@@ -1,1 +1,5 @@
-import './thumbnail.js';
+import { initThumbnail } from './thumbnail.js';
+import { getSimilarWizards } from './mock.js';
+
+const data = getSimilarWizards();
+initThumbnail(data);
