+++ js/thumbnail.js
@@ -1,18 +1,16 @@
-import { similarWizards } from './util.js';
-
 const picturesList = document.querySelector('.pictures');
 const pictureWizardTemplate = document.querySelector('#picture').content.querySelector('.picture');
 
-const createWizards = similarWizards();
-
 const createWizardsFragment = document.createDocumentFragment();
+const initThumbnail = (data) => {
+  data.forEach(({url, likes, comments}) => {
+    const wizardElement = pictureWizardTemplate.cloneNode(true);
+    wizardElement.querySelector('.picture__img').src = url;
+    wizardElement.querySelector('.picture__likes').textContent = likes;
+    wizardElement.querySelector('.picture__comments').textContent = comments.length;
+    createWizardsFragment.append(wizardElement);
+  });
+  picturesList.append(createWizardsFragment);
+};
 
-createWizards.forEach((wizard) => {
-  const wizardElement = pictureWizardTemplate.cloneNode(true);
-  wizardElement.querySelector('.picture__img').src = wizard.url;
-  wizardElement.querySelector('.picture__likes').textContent = wizard.likes;
-  wizardElement.querySelector('.picture__comments').textContent = wizard.comments.length;
-  createWizardsFragment.append(wizardElement);
-});
-
-picturesList.append(createWizardsFragment);
+export { initThumbnail };
